third_party/WebKit/LayoutTests/editing/selection/5497643.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<p>This tests to make sure that a selection inside a textarea is updated when the textarea is removed from the document.</p>
<textarea id="textarea"></textarea>
<div id="log"></div>
<script>
textarea = document.getElementById("textarea");
textarea.setSelectionRange(0, 0);
textarea.parentNode.removeChild(textarea);
test(function() {
    assert_equals(window.getSelection().type, 'Caret');
    assert_equals(window.getSelection().getRangeAt(0).startContainer, document.body);
    assert_equals(window.getSelection().getRangeAt(0).startOffset, 2);
});
</script>
